// src/app/core/data.d.ts
interface AppData {
    /** Preset definitions. */
    presets: Record<string, {
        /** List of active provider IDs in this preset. */
        providers: string[],
    }>;

    /** Currently active preset name. */
    currentPreset: string,

    /** Currently viewed URL. HTTPS protocol assumed. */
    currentUrl: string,
}

export interface ProviderData {
    /** Provider-specific data storage. */
    data: unknown,

    /**
     * Definition of item groups under this provider. For providers that
     * do not support grouping, this field is an empty object.
     *
     * The app manages item grouping for each provider and providers only
     * need to provide a flat list of items in `Provider.render()`.
     *
     * Each group under a provider must have a unique name, serving as the
     * identifier for the group within that provider. The "ungrouped" group
     * is represented by an empty string as the group name. Order of groups
     * is managed by the app and is defined by the order of this array.
     *
     * Each group contains a list of item identifiers that belong to that
     * group. An item can only belong to one group at a time.
     *
     * Items that are not listed in any group are considered ungrouped and
     * will be displayed in a default "ungrouped" group at top of all other
     * groups.
     *
     * Order of groups is defined by the `groupOrder` field below.
     *
     * Order of items within each group as well as in the ungrouped group
     * is determined by the `sortKey` field of each item. The order of the
     * `items` array of each group is not preserved.
     **/
    groups: Record<string, {
        /** List of item IDs contained in this group. */
        items: string[],
    }>,

    /** Order of groups under this provider. */
    groupOrder: string[],

    /** List of item IDs that are expanded in the UI. */
    expandedItems: string[],

    /** List of group names that are expanded in the UI. */
    expandedGroups: string[],
}

export interface Cache {
    providers: Record<string, unknown>,
}

/** The uniform interface for documentation providers. */
export interface Provider<T = unknown, TCache = unknown>
    extends ProviderInfo {
    /** Render a full view model from provider-specific data storage. */
    render(provider: ProviderContext<T, TCache>): ProviderOutput,
}

/** Metadata about a provider. */
export interface ProviderInfo {
    /** Unique identifier of the provider. */
    readonly id: string,

    /** Display name of the provider. */
    readonly name: string,

    /** Whether to enable item grouping for this provider. */
    readonly enableItemGrouping: boolean,

    /** Whether to render item names in <code> tags (monospace font). */
    readonly renderItemNameAsCode: boolean,
}

export interface ProviderContext<T = unknown, TCache = unknown> {
    /** Provider-specific data storage. */
    readonly data: T,

    /** Update the provider-specific data storage. */
    updateData(updater: (draft: T) => void): void,

    /** Provider-specific cache storage. */
    readonly cache: TCache,

    /** Update the provider-specific cache storage. */
    updateCache(updater: (draft: TCache) => void): void,

    /** The current URL being viewed in the app. HTTPS protocol assumed. */
    readonly currentUrl: string,

    /** Set the current URL being viewed in the app. HTTPS protocol assumed. */
    setCurrentUrl(url: string): void,
}

export type ProviderOutput = ReadonlyDeep<{
    items: Record<string, Item>,
    actions?: ProviderAction[],
}>;

export type ProviderAction =
    | { type: "node", render(): ReactNode }
    | { type: "menu", name: string, icon: IconProp, invoke(): void };

/** The uniform view model of a documentation item. */
export interface Item {
    /** Unique identifier for the item within its provider. */
    id: string,

    /** Display name of the item. */
    name: string,

    /** Sort key for ordering items within a group. */
    sortKey: string,

    /** List of documentation pages for this item. */
    pages: Page[],

    /** List of external links for this item. */
    links?: ItemLink[],

    /** List of custom actions for this item. */
    actions?: ItemAction[],

    /** For items that represents a package (or crate, module, etc.), this field
     *  contains the view model of the version selector. **/
    versions?: ItemVersions,
}